HWR_ProcessSeg: fix cliphigh overflow

+++ src/hardware/hw_main.cpp
@@ -1352,7 +1352,7 @@ void HWR_ProcessSeg(void) // Sort of like GLWall::Process in GZDoom
 	// x offset the texture
 	const fixed_t texturehpeg = gl_sidedef->textureoffset + gl_curline->offset;
 	const float cliplow  = (float)texturehpeg;
-	const float cliphigh = (float)(texturehpeg + gl_curline->length);
+	const float cliphigh = ((float)texturehpeg + (float)gl_curline->length);
 
 	FUINT lightnum = gl_frontsector->lightlevel;
 	extracolormap_t *colormap = gl_frontsector->extra_colormap;
